Change names of packages Add request and response mappings
