--- src/main/java/com/haanido/playercreator/PlayerCreatorApplication.java
@@ -1,13 +1,0 @@
-package com.haanido.playercreator;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PlayerCreatorApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PlayerCreatorApplication.class, args);
-	}
-
-}
