--- src/main/java/com/haanido/playercreator/entity/Player.java
@@ -1,38 +1,0 @@
-package com.haanido.playercreator.entity;
-
-import jakarta.persistence.*;
-import lombok.Data;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Table
-@Data
-public class Player {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private int id;
-    @Column(name = "first_name")
-    private String firstName;
-    @Column(name = "last_name")
-    private String lastName;
-    private String email;
-    private int age;
-    private double ranking;
-    @OneToOne(cascade = CascadeType.ALL)
-    @JoinColumn(name = "address_id")
-    private Address address;
-    @OneToMany(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-    @JoinColumn(name = "player_id")
-    private List<PhoneNumber> phoneNumbers;
-
-    public void addPhoneNumber(PhoneNumber phoneNumber) {
-        if (phoneNumbers == null) {
-            phoneNumbers = new ArrayList<>();
-        }
-        phoneNumbers.add(phoneNumber);
-    }
-}
