--- src/main/java/com/haanido/playercreator/repo/AddressRepository.java
@@ -1,7 +1,0 @@
-package com.haanido.playercreator.repo;
-
-import com.haanido.playercreator.entity.Address;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AddressRepository extends JpaRepository<Address, Integer> {
-}
