--- src/main/java/com/haanido/playercreator/repo/PhoneNumberRepository.java
@@ -1,9 +1,0 @@
-package com.haanido.playercreator.repo;
-
-import com.haanido.playercreator.entity.PhoneNumber;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PhoneNumberRepository extends JpaRepository<PhoneNumber, Integer> {
-}
