--- src/main/java/com/haanido/playercreator/rest/PlayerNotFoundException.java
@@ -1,15 +1,0 @@
-package com.haanido.playercreator.rest;
-
-public class PlayerNotFoundException extends RuntimeException{
-    public PlayerNotFoundException(String message) {
-        super(message);
-    }
-
-    public PlayerNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public PlayerNotFoundException(Throwable cause) {
-        super(cause);
-    }
-}
