--- src/main/java/com/haanido/playercreator/service/AddressService.java
@@ -1,16 +1,0 @@
-package com.haanido.playercreator.service;
-
-import com.haanido.playercreator.repo.AddressRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class AddressService {
-
-    private final AddressRepository addressRepository;
-
-    public void deleteAddress(int id) {
-        addressRepository.deleteById(id);
-    }
-}
