+++ src/main/java/com/sogeti/playerrestapi/PlayerRestApiApplication.java
@@ -1,0 +1,20 @@
+package com.sogeti.playerrestapi;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+
+@SpringBootApplication
+public class PlayerRestApiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PlayerRestApiApplication.class, args);
+	}
+
+	@Bean
+	public ModelMapper modelMapper() {
+		return new ModelMapper();
+	}
+
+}
