+++ src/main/java/com/sogeti/playerrestapi/entity/PhoneNumber.java
@@ -1,0 +1,15 @@
+package com.sogeti.playerrestapi.entity;
+
+import jakarta.persistence.*;
+import lombok.Data;
+
+@Entity
+@Table(name = "phonenumber")
+@Data
+public class PhoneNumber {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+    private String number;
+    private PhoneNumberType type;
+}
