+++ src/main/java/com/sogeti/playerrestapi/entity/PhoneNumberType.java
@@ -1,0 +1,17 @@
+package com.sogeti.playerrestapi.entity;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum PhoneNumberType {
+
+    HOME("H"),
+    WORK("W"),
+    MOBILE("M");
+
+    private final String code;
+
+
+}
