+++ src/main/java/com/sogeti/playerrestapi/model/AddressResponse.java
@@ -1,0 +1,15 @@
+package com.sogeti.playerrestapi.model;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+
+@Data
+@NoArgsConstructor
+public class AddressResponse {
+    private String street;
+    private int houseNumber;
+    private String zipcode;
+    private String city;
+    private String countryCode;
+}
