+++ src/main/java/com/sogeti/playerrestapi/model/PlayerAddRequest.java
@@ -1,0 +1,15 @@
+package com.sogeti.playerrestapi.model;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class PlayerAddRequest {
+
+    private String firstName;
+    private String lastName;
+    private String email;
+    private int age;
+    private double ranking;
+}
