+++ src/main/java/com/sogeti/playerrestapi/repo/AddressRepository.java
@@ -1,0 +1,7 @@
+package com.sogeti.playerrestapi.repo;
+
+import com.sogeti.playerrestapi.entity.Address;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AddressRepository extends JpaRepository<Address, Integer> {
+}
