+++ src/main/java/com/sogeti/playerrestapi/rest/PlayerErrorResponse.java
@@ -1,0 +1,12 @@
+package com.sogeti.playerrestapi.rest;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class PlayerErrorResponse {
+    private int status;
+    private String message;
+    private long timestamp;
+}
