+++ src/main/java/com/sogeti/playerrestapi/rest/PlayerNotFoundException.java
@@ -1,0 +1,15 @@
+package com.sogeti.playerrestapi.rest;
+
+public class PlayerNotFoundException extends RuntimeException{
+    public PlayerNotFoundException(String message) {
+        super(message);
+    }
+
+    public PlayerNotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public PlayerNotFoundException(Throwable cause) {
+        super(cause);
+    }
+}
